# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DSA Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /topics/{topicId}/progress:
    patch:
      operationId: updateTopicProgress
      tags: [topics]
      summary: Update user progress for a topic
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicProgressInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicWithProgress"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Questions ─────────────────────────────────────────────────────────────
...
  /revisions:
    get:
      operationId: listRevisions
      tags: [revisions]
      summary: List revisions (optionally filter by status/due)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Pending, Completed]
        - name: dueToday
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of revisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevisionWithDetails"
  /revisions/{revisionId}/complete:
    patch:
      operationId: completeRevision
      tags: [revisions]
      summary: Mark a revision as completed
      parameters:
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Notes ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TopicWithProgress:
      type: object
      required: [id, name, description, status, questionCount, solvedCount]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Not Started, In Progress, Completed]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        questionCount:
          type: integer
        solvedCount:
          type: integer
    TopicProgressInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [Not Started, In Progress, Completed]

    # ── Questions ─────────────────────────────────────────────────────────
...
    RevisionWithDetails:
      type: object
      required: [id, revisionDate, status, revisionNumber]
      properties:
        id:
          type: integer
        revisionDate:
          type: string
        status:
          type: string
          enum: [Pending, Completed]
        revisionNumber:
          type: integer
        topicId:
          type: ["integer", "null"]
        topicName:
          type: ["string", "null"]
        questionId:
          type: ["integer", "null"]
        questionTitle:
          type: ["string", "null"]
        questionDifficulty:
          type: ["string", "null"]

    # ── Notes ─────────────────────────────────────────────────────────────
...